% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debt_analysis.R
\name{debt_analysis}
\alias{debt_analysis}
\title{Conducts linear regression and creates a linear plot for Public Debt data from FRED}
\usage{
debt_analysis(data)
}
\arguments{
\item{data}{The Public Debt dataset from FRED}
}
\value{
A summary of the linear regression as well as plot with Public Debt over time
}
\description{
Conducts linear regression and creates a linear plot for Public Debt data from FRED
}
