% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debt_gdp_analysis.R
\name{debt_gdp_analysis}
\alias{debt_gdp_analysis}
\title{Combines the GDP and Public Dataset from FRED and plots both on a single graph}
\usage{
debt_gdp_analysis(debt, gdp)
}
\arguments{
\item{debt}{The Public Debt dataset from FRED}

\item{gdp}{The GDP dataset from FRED}
}
\value{
A plot of Public Debt (in thousands) and GDP (in billions) on the same graph
}
\description{
Combines the GDP and Public Dataset from FRED and plots both on a single graph
}
