% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Econanalysis-package.R
\docType{package}
\name{Econanalysis-package}
\alias{Econanalysis}
\alias{Econanalysis-package}
\title{Econanalysis: Conduct Analysis on Datasets From FRED}
\description{
This package is used to conduct analysis on datasets from FRED. It is used to automate tasks like making certain plots and conducting linear regression as well as transforming datasets for otheer use.
}
\author{
\strong{Maintainer}: Tanvi Chiddarwar \email{tanvi.chiddarwar@gmail.com}

}
\keyword{internal}
