% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdp_analysis.R
\name{gdp_analysis}
\alias{gdp_analysis}
\title{Conducts linear regression and creates a linear plot for GDP data from FRED}
\usage{
gdp_analysis(data)
}
\arguments{
\item{data}{The GDP dataset from FRED}
}
\value{
A summary of the linear regression as well as plot with GDP over time
}
\description{
Conducts linear regression and creates a linear plot for GDP data from FRED
}
