% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/okuns_law.R
\name{okuns_law}
\alias{okuns_law}
\title{Conduct Analysis of the relationship of percent change of GDP and Unemployment: Okun's Law}
\usage{
okuns_law(gdp, unemployment)
}
\arguments{
\item{gdp}{The GDP dataset from FRED}

\item{unemployment}{The unemployment rate dataset from FRED}
}
\value{
Summary of linear regression of the relationship between percent change in GDP and unemployment as well as a plot depicting this relationship
}
\description{
Conduct Analysis of the relationship of percent change of GDP and Unemployment: Okun's Law
}
