% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent_change.R
\name{percent_change}
\alias{percent_change}
\title{Makes a new column that has percent change from year to year}
\usage{
percent_change(data, var)
}
\arguments{
\item{data}{The dataset to modify}

\item{var}{The variable that to be put in terms of percent change}
}
\value{
A new dataset with the percent change column
}
\description{
Makes a new column that has percent change from year to year
}
\examples{
# Data: unemp

#Var: UNRATE
unemp <- percent_change(unemp, UNRATE)

}
