% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/philips.R
\name{philips}
\alias{philips}
\title{Conduct Analysis of the relationship between inflation and unemployment: Philip's Curve}
\usage{
philips(infl, unemp)
}
\arguments{
\item{infl}{The inflation rate dataset from FRED}

\item{unemp}{The unemployment rate dataset from FRED}
}
\value{
A plot depicting the relationship between inflation rate and unemployment
}
\description{
Conduct Analysis of the relationship between inflation and unemployment: Philip's Curve
}
